--- efesto/efesto-core/efesto-identifiers-api-decisions/src/main/java/org/kie/efesto/identifiers/api/decisions/DecisionServiceIds.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.efesto.identifiers.api.decisions;
-
-public class DecisionServiceIds {
-    private final LocalDecisionId parentId;
-
-    public DecisionServiceIds(LocalDecisionId parentId) {
-        this.parentId = parentId;
-    }
-
-    public LocalDecisionServiceId get(String serviceId) {
-        return new LocalDecisionServiceId(parentId, serviceId);
-    }
-}
