--- efesto/efesto-core/efesto-identifiers-api-decisions/src/main/java/org/kie/efesto/identifiers/api/decisions/LocalDecisionServiceId.java
@@ -1,52 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.efesto.identifiers.api.decisions;
-
-import org.kie.efesto.identifiers.api.application.Id;
-import org.kie.efesto.identifiers.api.application.LocalId;
-import org.kie.efesto.identifiers.api.application.LocalUriId;
-
-public class LocalDecisionServiceId extends LocalUriId implements LocalId {
-    public static final String PREFIX = "services";
-
-    private final Id decisionId;
-    private final String serviceId;
-
-    public LocalDecisionServiceId(Id decisionId, String serviceId) {
-        super(decisionId.toLocalId().asLocalUri().append(PREFIX).append(serviceId));
-        LocalId localDecisionId = decisionId.toLocalId();
-        if (!localDecisionId.toLocalId().asLocalUri().startsWith(LocalDecisionId.PREFIX)) {
-            throw new IllegalArgumentException("Not a valid decision path"); // fixme use typed exception
-        }
-        this.decisionId = decisionId;
-        this.serviceId = serviceId;
-    }
-
-    public Id decisionId() {
-        return decisionId;
-    }
-
-    public String serviceId() {
-        return serviceId;
-    }
-
-    @Override
-    public LocalId toLocalId() {
-        return this;
-    }
-
-}
