--- efesto/efesto-core/efesto-identifiers-api-predictions/src/main/java/org/kie/efesto/identifiers/api/predictions/LocalPredictionId.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.efesto.identifiers.api.predictions;
-
-import org.kie.efesto.identifiers.api.application.Id;
-import org.kie.efesto.identifiers.api.application.LocalId;
-import org.kie.efesto.identifiers.api.application.LocalUri;
-import org.kie.efesto.identifiers.api.application.LocalUriId;
-
-public class LocalPredictionId extends LocalUriId implements Id {
-    public static final String PREFIX = "predictions";
-
-    private final String name;
-
-    public LocalPredictionId(String name) {
-        super(LocalUri.Root.append(PREFIX).append(name));
-        this.name = name;
-    }
-
-    public String name() {
-        return name;
-    }
-
-    @Override
-    public LocalId toLocalId() {
-        return this;
-    }
-
-}
