--- efesto/efesto-core/efesto-identifiers-api-rules/pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2021 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.kie</groupId>
-        <artifactId>efesto-core</artifactId>
-        <version>8.24.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>efesto-identifiers-api-rules</artifactId>
-    <name>Efesto :: Identifiers :: Api :: Rules</name>
-
-    <properties>
-        <java.module.name>org.kie.efesto.identifiers.api.rules</java.module.name>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.kie</groupId>
-            <artifactId>efesto-identifiers-api-application</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-api</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-</project>
